Hardware Serial for Teensy 4
First pass update for Hardware Serial for Teensy4

Currently using one Class for all of the objects.

Now using ISR for Reads and writes.

Still lots of work left, like filling in some
of the other features like the RS485 support

Moving each of the objects to their own files as to not
bring in code and data if object not used.

Yield stuff, half duplex.

Not sure yet if some of the other pins like TX, or RTS, CTS
need to have their select input registers set as well
...

But a start