I Hope fix MTP Disk and Serial Emulation
We were not getting any Serial Emulated output when using the Experimental MTP Disk setup.

I believe the issue is the athe MTP_INTERFACE number and the SEREUM_INTERFACE number need to match the order the two interfaces are defined in the usb_desc.c file.
And in this case SEREMU is defined before MTP_INTERFACE.

Fix is to change the two numbers.

Alternative would be move the MTP_DISK define up in the order in which all of the descriptors are written out.