T3/T4.x MTP device descriptor - update endpoint and size
For T4.x to allow MTP USB type device to generate events to send to the host, the Endpoint configuration need to change to say that we are transmitting on interrupt, not receiving.

Plus I updated the end point event size to 32 as some events might be > 16.  16 is only large enought so send one parameter.  I noticed Kindle Fire has it set to something like 28...

I have done similar changes for T3.x but have not tested it yet.