RAW HID optional support on T4.x for 512 byte transfers
While testing different possible things that may be causing issues in the FS integration work, I thought I would try to emulate the MTP transfer of 512 bytes.  And as the MTP code is modeled after RAWHID, thought it would be interesting to experiment with 512 byte transfers.

I setup the code such that on T4.x you can build for either 64 byte or 512 bytes.

Added methods to query to see what size the transfers are.

Again not sure if this is going anywhere, although it does transfer data 8 times faster for obvious reasons.